pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agentvault"
version = "0.1.0"
description = "Budget controller open source pour agents IA dépensant de l'USDC sur Base"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["ai-agents", "usdc", "budget", "blockchain", "base", "web3"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "web3>=7.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/BigFoot3/agentvault"
Repository = "https://github.com/BigFoot3/agentvault"
Issues = "https://github.com/BigFoot3/agentvault/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentvault*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["agentvault"]
omit = ["tests/*", "venv/*"]
